fix: execute `PreBlocker` for upgrade module (#707)
