+++ app/upgrades/vanilla/upgrades_test.go
@@ -1,73 +1,25 @@
-package upgrades_test
+package vanilla_test
 
 import (
-	"context"
 	"fmt"
-	"math/rand"
 	"testing"
 	"time"
 
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/header"
-	store "cosmossdk.io/store/types"
 	"cosmossdk.io/x/upgrade"
 	upgradetypes "cosmossdk.io/x/upgrade/types"
 	"github.com/babylonchain/babylon/app"
-	"github.com/babylonchain/babylon/app/keepers"
-	"github.com/babylonchain/babylon/app/upgrades"
-	"github.com/babylonchain/babylon/testutil/datagen"
-	btcstakingkeeper "github.com/babylonchain/babylon/x/btcstaking/keeper"
+	v1 "github.com/babylonchain/babylon/app/upgrades/vanilla"
 	bstypes "github.com/babylonchain/babylon/x/btcstaking/types"
 	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
 	"github.com/stretchr/testify/suite"
 )
 
 const (
 	DummyUpgradeHeight = 5
 )
-
-var Upgrade = upgrades.Upgrade{
-	UpgradeName:          "vanilla",
-	CreateUpgradeHandler: CreateUpgradeHandler,
-	StoreUpgrades:        store.StoreUpgrades{},
-}
-
-func CreateUpgradeHandler(
-	mm *module.Manager,
-	_ module.Configurator,
-	_ upgrades.BaseAppParamManager,
-	keepers *keepers.AppKeepers,
-) upgradetypes.UpgradeHandler {
-	return func(context context.Context, _plan upgradetypes.Plan, vm module.VersionMap) (module.VersionMap, error) {
-		ctx := sdk.UnwrapSDKContext(context)
-
-		propVanilla(ctx, &keepers.AccountKeeper, &keepers.BTCStakingKeeper)
-
-		return vm, nil
-	}
-}
-
-func propVanilla(
-	ctx sdk.Context,
-	accountKeeper *authkeeper.AccountKeeper,
-	bsKeeper *btcstakingkeeper.Keeper,
-) {
-	r := rand.New(rand.NewSource(time.Now().Unix()))
-
-	// remove an account
-	allAccounts := accountKeeper.GetAllAccounts(ctx)
-	accountKeeper.RemoveAccount(ctx, allAccounts[len(allAccounts)-1])
-
-	// insert a FP
-	fp, err := datagen.GenRandomFinalityProvider(r)
-	if err != nil {
-		panic(err)
-	}
-	bsKeeper.SetFinalityProvider(ctx, fp)
-}
 
 type UpgradeTestSuite struct {
 	suite.Suite
@@ -78,6 +30,10 @@ type UpgradeTestSuite struct {
 }
 
 func (s *UpgradeTestSuite) SetupTest() {
+	// add the upgrade plan
+	app.Upgrades = append(app.Upgrades, v1.Upgrade)
+
+	// set up app
 	s.app = app.Setup(s.T(), false)
 	s.ctx = s.app.BaseApp.NewContextLegacy(false, tmproto.Header{Height: 1, ChainID: "babylon-1", Time: time.Now().UTC()})
 	s.preModule = upgrade.NewAppModule(s.app.UpgradeKeeper, s.app.AccountKeeper.AddressCodec())
@@ -106,23 +62,16 @@ func (s *UpgradeTestSuite) TestUpgradePayments() {
 			func() {
 				// inject upgrade plan
 				s.ctx = s.ctx.WithBlockHeight(DummyUpgradeHeight - 1)
-				plan := upgradetypes.Plan{Name: Upgrade.UpgradeName, Height: DummyUpgradeHeight}
-				s.app.UpgradeKeeper.SetUpgradeHandler(
-					Upgrade.UpgradeName,
-					Upgrade.CreateUpgradeHandler(
-						s.app.ModuleManager,
-						nil,
-						nil,
-						s.app.AppKeepers,
-					),
-				)
-
-				// run upgrade
+				plan := upgradetypes.Plan{Name: v1.Upgrade.UpgradeName, Height: DummyUpgradeHeight}
 				err := s.app.UpgradeKeeper.ScheduleUpgrade(s.ctx, plan)
 				s.NoError(err)
-				_, err = s.app.UpgradeKeeper.GetUpgradePlan(s.ctx)
-				s.NoError(err)
 
+				// ensure upgrade plan exists
+				actualPlan, err := s.app.UpgradeKeeper.GetUpgradePlan(s.ctx)
+				s.NoError(err)
+				s.Equal(plan, actualPlan)
+
+				// execute upgrade
 				s.ctx = s.ctx.WithHeaderInfo(header.Info{Height: DummyUpgradeHeight, Time: s.ctx.BlockTime().Add(time.Second)}).WithBlockHeight(DummyUpgradeHeight)
 				s.NotPanics(func() {
 					_, err := s.preModule.PreBlock(s.ctx)
